% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation_analysis.R
\name{calculation_analysis}
\alias{calculation_analysis}
\title{Model run calculation}
\usage{
calculation_analysis(data, time1, time2, time3, status, covs, K)
}
\arguments{
\item{data}{and variables and user input}
}
\value{
A complete output object
}
\description{
This function runs the model and estimates the spline coefficients,
regression coefficients, and their covariance matrix.
}
