% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gicsurv.R
\name{gicsurv}
\alias{gicsurv}
\title{app function}
\usage{
gicsurv(...)
}
\arguments{
\item{NULL}{once run, the shiny app will be prompted}
}
\value{
Estimates of covariate effects and baseline estimates and plots
}
\description{
This function is to run the shiny app that will estimate covariate effects
and tabulate and visualize estimates of the baseline hazard and survival.
}
