% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSE.R
\name{importSE}
\alias{importSE}
\title{import server}
\usage{
importSE(id)
}
\arguments{
\item{session}{id}
}
\value{
import server objects
}
\description{
This function imports the data and renders the data into a datatable.
}
