% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importUI.R
\name{importUI}
\alias{importUI}
\title{data import UI}
\usage{
importUI(id)
}
\arguments{
\item{session}{number}
}
\value{
The UI objects for the data import UI
}
\description{
This function is the for creating the UI of the data import for the current
session.
}
