% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_analysis.R
\name{output_analysis}
\alias{output_analysis}
\title{coefficient estimates calculator}
\usage{
output_analysis(data, results, covs, K)
}
\arguments{
\item{data}{model object and covariates}
}
\value{
covariate effects
}
\description{
This function calculates estimates of the covariate effects given the data
and selected variables and after running of the estimation.
}
