% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_analysis.R
\name{pred_analysis}
\alias{pred_analysis}
\title{baseline estimates calculator}
\usage{
pred_analysis(
  results,
  data,
  time1,
  time2,
  time3,
  status,
  covs,
  maxtime,
  nboot,
  K
)
}
\arguments{
\item{data}{and variables and user input}
}
\value{
Baseline estimates
}
\description{
This function calculates baseline hazard, cumulative hazard and survival
estimates given the data and selected variables and after running of the
estimation.
}
