% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varselect_server.R
\name{varselect_server}
\alias{varselect_server}
\title{variable server}
\usage{
varselect_server(id, data)
}
\arguments{
\item{session}{number and data}
}
\value{
updated variable selection
}
\description{
This function builds the server object to update the variable selection
depending on the uploaded data.
}
