% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varselect_ui.R
\name{varselect_ui}
\alias{varselect_ui}
\title{variable select UI}
\usage{
varselect_ui(id)
}
\arguments{
\item{session}{number}
}
\value{
The UI objects for the variable selection
}
\description{
This function builds the UI for the variable selection.
}
