% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warning_calc.R
\name{warning_calc}
\alias{warning_calc}
\title{warning indicator}
\usage{
warning_calc(data, time1, time2, time3, status)
}
\arguments{
\item{data}{file and selected variables}
}
\value{
An indicator
}
\description{
This function allows a pop-up if the variables have not been changed from the
default.
}
